+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming version
 
+- Fixed companion bags on Android (#116)
 - Fixed unexpected move of recruited spouse in FarmHouse.
 - Player can gain extra friendship points every whole hour while adventuring with companions (#109)
 - Mod's translation was moved to the separated folder `/locale/<localeCode>` and not requires the suffix ~~`.<localeCode>.json`~~